+++ src/components/Formik/index.js
@@ -1,7 +1,9 @@
 import FormikInputField from './FormikInputField';
 import FormikSelectField from './FormikSelectField';
+import FormikTextEditorField from './FormikTextEditorField';
 
 export {
+  FormikTextEditorField,
   FormikInputField,
   FormikSelectField,
 };