+++ src/config/menu.js
@@ -92,6 +92,10 @@ const sidebarTopMenu = [{
     label: 'SIDEBAR.TOP_MENU.BRAND_CONFIG_CREATE',
     url: '/brand-config/create',
     permissions: new Permissions(permissions.BRAND_CONFIG.CREATE_BRAND_CONFIG),
+  }, {
+    label: 'SIDEBAR.TOP_MENU.EMAIL_TEMPLATES',
+    url: '/email-templates',
+    permissions: new Permissions(permissions.EMAIL_TEMPLATES.CREATE_EMAIL_TEMPLATE),
   }],
 }];