+++ src/routes/IndexRoute/IndexRoute.jsx
@@ -28,6 +28,7 @@ import Callbacks from 'routes/Callbacks';
 import ReleaseNotes from 'routes/ReleaseNotes';
 import PersonalDashboard from 'routes/PersonalDashboard';
 import BrandConfig from 'routes/BrandConfig';
+import EmailTemplates from 'routes/EmailTemplates';
 import { operatorsExcludeAuthorities } from 'config/menu';
 import { withStorage } from 'providers/StorageProvider';
 
@@ -97,6 +98,7 @@ class IndexRoute extends PureComponent {
             <Route path="/sales-rules" layout={MainLayout} component={SalesRules} isPrivate />
             <Route path="/release-notes" layout={MainLayout} component={ReleaseNotes} isPrivate />
             <Route path="/brand-config" layout={MainLayout} component={BrandConfig} isPrivate />
+            <Route path="/email-templates" layout={MainLayout} component={EmailTemplates} isPrivate />
             <Route path="/logout" component={Logout} isPrivate />
 
             {/* Public routes */}